requirements.txt
Flask
logging
matplotlib
pandas_datareader
python-dateutil
scipy
setuptools
yfinance